Add: #231 Wave 2 세션 F — feature/home/impl KMP+CMP 전환 (소스 레이아웃 재편까지, UI commonMain 이전은 후속)
