+++ feature/home/impl/src/androidMain/kotlin/me/pecos/memozy/presentation/components/BottomNavBar.kt
@@ -1,0 +1,70 @@
+package me.pecos.memozy.presentation.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import dev.chrisbanes.haze.HazeState
+import dev.chrisbanes.haze.HazeStyle
+import me.pecos.memozy.presentation.theme.LocalAppColors
+
+// ── 바텀 네비게이션 ─────────────────────────────────────────────────────────────
+
+@Composable
+fun FloatingNavPill(
+    selectedRoute: String,
+    onItemSelected: (String) -> Unit,
+    hazeState: HazeState,
+    glassStyle: HazeStyle,
+    modifier: Modifier = Modifier
+) {
+    val colors = LocalAppColors.current
+    Row(
+        modifier = modifier
+            .height(52.dp)
+            .clip(RoundedCornerShape(50))
+            .background(colors.navBackground.copy(alpha = 0.75f))
+            .border(width = 1.dp, color = colors.navBorder, shape = RoundedCornerShape(50))
+            .padding(horizontal = 2.dp, vertical = 2.dp),
+        horizontalArrangement = Arrangement.SpaceEvenly,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        bottomNavItems.forEach { item ->
+            val selected = selectedRoute == item.route
+            Box(
+                modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight()
+                    .clip(RoundedCornerShape(50))
+                    .background(
+                        if (selected) colors.navIconSelected.copy(alpha = 0.06f)
+                        else Color.Transparent
+                    )
+                    .clickable { onItemSelected(item.route) },
+                contentAlignment = Alignment.Center
+            ) {
+                Icon(
+                    imageVector = item.icon,
+                    contentDescription = item.label,
+                    modifier = Modifier.size(24.dp),
+                    tint = if (selected) colors.navIconSelected else colors.navIconUnselected
+                )
+            }
+        }
+    }
+}
+
