--- feature/home/impl/src/main/java/me/pecos/memozy/presentation/components/BottomNavBar.kt
@@ -1,139 +1,0 @@
-package me.pecos.memozy.presentation.components
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.wanted.android.wanted.design.theme.DesignSystemTheme
-import dev.chrisbanes.haze.HazeState
-import dev.chrisbanes.haze.HazeStyle
-import dev.chrisbanes.haze.HazeTint
-import dev.chrisbanes.haze.hazeEffect
-import dev.chrisbanes.haze.rememberHazeState
-import me.pecos.memozy.feature.core.resource.R as ResourceR
-import me.pecos.memozy.presentation.theme.LocalAppColors
-
-// ── 바텀 네비게이션 ─────────────────────────────────────────────────────────────
-
-@Composable
-fun FloatingNavPill(
-    selectedRoute: String,
-    onItemSelected: (String) -> Unit,
-    hazeState: HazeState,
-    glassStyle: HazeStyle,
-    modifier: Modifier = Modifier
-) {
-    val colors = LocalAppColors.current
-    Row(
-        modifier = modifier
-            .height(52.dp)
-            .clip(RoundedCornerShape(50))
-            .background(colors.navBackground.copy(alpha = 0.75f))
-            .border(width = 1.dp, color = colors.navBorder, shape = RoundedCornerShape(50))
-            .padding(horizontal = 2.dp, vertical = 2.dp),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        bottomNavItems.forEach { item ->
-            val selected = selectedRoute == item.route
-            Box(
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxHeight()
-                    .clip(RoundedCornerShape(50))
-                    .background(
-                        if (selected) colors.navIconSelected.copy(alpha = 0.06f)
-                        else Color.Transparent
-                    )
-                    .clickable { onItemSelected(item.route) },
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    imageVector = item.icon,
-                    contentDescription = item.label,
-                    modifier = Modifier.size(24.dp),
-                    tint = if (selected) colors.navIconSelected else colors.navIconUnselected
-                )
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true, backgroundColor = 0xFFF0F0F0)
-@Composable
-fun FloatingNavPillPreview() {
-    DesignSystemTheme {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(120.dp)
-                .background(Color(0xFFF0F0F0)),
-            contentAlignment = Alignment.BottomCenter
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp, vertical = 12.dp)
-                    .height(IntrinsicSize.Max),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                FloatingNavPill(
-                    selectedRoute = "main",
-                    onItemSelected = {},
-                    hazeState = rememberHazeState(),
-                    glassStyle = HazeStyle(
-                        blurRadius = 20.dp,
-                        backgroundColor = Color.White,
-                        tints = listOf(HazeTint(color = Color.White.copy(alpha = 0.4f)))
-                    ),
-                    modifier = Modifier.weight(1f)
-                )
-                Spacer(modifier = Modifier.width(12.dp))
-                Box(
-                    modifier = Modifier
-                        .fillMaxHeight()
-                        .aspectRatio(1f)
-                        .shadow(10.dp, CircleShape)
-                        .background(Color.White.copy(alpha = 0.88f), CircleShape)
-                        .border(0.5.dp, Color.White.copy(alpha = 0.7f), CircleShape)
-                        .clip(CircleShape),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Image(
-                        painter = painterResource(id = ResourceR.drawable.icon_add),
-                        contentDescription = "메모 추가",
-                        colorFilter = ColorFilter.tint(Color.Black),
-                        modifier = Modifier.size(42.dp)
-                    )
-                }
-            }
-        }
-    }
-}
